Add files via upload
The full solver procedure is contain within the file rheoFoamLC.C. The header file createFields.H declares/initialises the relevant fields and variables: velocity, pressure, microstructure and stress. The latter two fields are specified through the constitutiveModel2 object; in case of the Leslie-Ericksen model the microstructure is described through the director field, while in tensorial models, the microstructure is described through the Q-tensor. Finally, the specific form of the constitutive equation is contained in the object LE_1constant, where the director evolution equation and stress definition are specified.

the file homeotropic describes the implementation of a homeotropic boundary condition.  We use the patch( ).nf( ) function, which computes the wall-normal vector and prescribes it on the wall. The procedure is not so straightforward in the wall-parallel case, where it is necessary to prescribe a geometry-specific function that generates a wall-parallel director orientation.